pyproject.toml
[project]
name = "scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite",
    "beautifulsoup4>=4.14.3",
    "greenlet>=3.3.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-telegram-bot[job-queue]>=22.6",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "sqlalchemy",
    "types-beautifulsoup4>=4.12.0.20250516",
    "types-requests>=2.32.4.20260107",
]

[dependency-groups]
dev = [
    "complexipy>=5.2.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.14",
    "ty",
]

[tool.pytest.ini_options]
pythonpath = "."
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
omit = [
    "get_chat_id.py",
    "tests/*"
]
